+++ metadata.rb
@@ -1,14 +1,23 @@
+# Copyright 2015 Nordstrom, Inc.
 #
-# Copyright (c) 2015 Nordstrom, Inc.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# Originally created from Nordstrom Cookbook Template v0.16.2
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 #
 
 name             'chef_vault_try_notify'
 maintainer       'Nordstrom, Inc.'
 maintainer_email 'techcheftm@nordstrom.com'
-license          'MIT'
+license          'Apache 2.0'
 description      'test if chef-vault items can be decrypted at runtime'
-version          '0.3.0'
+version          '0.4.0'
 
 depends          'chef-vault', '~> 1.3'