+++ spec/cookbooks/secret_waiter/decrypt_failure_spec.rb
@@ -1,3 +1,17 @@
+# Copyright 2015 Nordstrom, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 RSpec.describe 'secret_waiter::decrypt_failure' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new(
@@ -5,29 +19,20 @@ RSpec.describe 'secret_waiter::decrypt_failure' do
     ).converge(described_recipe)
   end
 
-  it 'calls the on_failure block if the secrets canot be decrypted' do
-    bag = double 'data bag'
-    %w(bar gzonk).each do |itemname|
-      expect(bag).to receive(:key?)
-        .at_least(:once)
-        .with("#{itemname}_keys")
-        .and_return(true)
-      expect(bag)
-        .to receive(:[])
-        .with("#{itemname}_keys")
-        .at_least(:once)
-        .and_raise(ChefVault::Exceptions::SecretDecryption)
-    end
-    %w(foo baz).each do |bagname|
-      expect(Chef::DataBag)
-        .to receive(:load)
-        .at_least(:once)
-        .with(bagname)
-        .and_return(bag)
-    end
+  it 'calls the on_failure block if passwords_data_bag/admin_data_bag_item cannot be decrypted' do
+    allow(ChefVault::Item)
+      .to receive(:vault?)
+      .with('passwords_data_bag', 'admin_data_bag_item')
+      .and_return(true)
+
+    allow(ChefVault::Item)
+      .to receive(:load)
+      .with('passwords_data_bag', 'admin_data_bag_item')
+      .and_raise(ChefVault::Exceptions::SecretDecryption)
+
     expect(Chef::Log)
       .to receive(:warn)
-      .at_least(:once)
+      .twice # because we're retrying 2x, each failure gets a chance to notify
       .with(/this is where we would send an SNS notification/)
     expect(Chef::Log)
       .to receive(:warn)