+++ test/fixtures/cookbooks/secret_waiter/recipes/decrypt_failure.rb
@@ -1,7 +1,21 @@
+# Copyright 2015 Nordstrom, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 include_recipe 'chef_vault_try_notify'
 
 chef_vault_try_notify 'my_secrets' do
-  vault_items %w(foo/bar baz/gzonk)
+  vault_items %w(passwords_data_bag/admin_data_bag_item)
   max_tries 2
   wait_period 1
   on_failure do |state|