+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Revision History for chef_vault_try_notify
 
+## 0.4.0
+
+* replace 'include ChefVaultItem' with 'include ChefVaultCookbook' to reflect
+change in chef-vault cookbook 1.3.1
+* fix specs
+* Switch from MIT to Apache 2.0 license
+
 ## 0.3.0
 
 * fix the predicate use of the guard state helper (it could throw a LocalJumpError exception previously)